pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "speedy-openai"
version = "0.1.0"
description = "Async OpenAI client for fast and efficient API requests using AIOHTTP module."
authors = ["Luca Ferrario <lucaferrario199@gmail.com>"]
readme = "README.md"
include = ["README.md", "CHANGELOG.md"]
license = "MIT"
keywords = ["openai", "async", "aiohttp", "api"]
homepage = "https://github.com/lucafirefox/speedy-openai"
repository = "https://github.com/lucafirefox/speedy-openai"
documentation = "https://github.com/lucafirefox/speedy-openai#readme"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
]
packages = [{ include = "speedy_openai" }]

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
loguru = "^0.7.3"
openai = "^1.57.4"
tiktoken = "^0.8.0"
tqdm = "^4.67.1"
tenacity = "^9.0.0"
aiohttp = "^3.11.10"


[tool.commitizen]
name = "cz_conventional_commits"
version = "1.3.1"
tag_format = "v$version"
version_files = [
    "pyproject.toml:^version",
    "speedy_openai/__init__.py:^__version__",
    ]
update_changelog_on_bump = true
changelog_incremental = true
bump_message = "bump: $current_version → $new_version"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
commitizen = "^3.29.0"
pytest-cov = "^6.0.0"
moto = {extras = ["s3"], version = "^5.0.22"}
pytest-asyncio = "^0.24.0"

